# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PYQDeck API Explorer",
    "version": "1.0.0",
    "description": "\nWelcome to the official **PYQDeck API Documentation**.\n\nThis API powers the PYQDeck platform, enabling users to access, bookmark, and study Previous Year Questions (PYQs) efficiently.\n\n### 🔑 Authentication\nMost endpoints require a valid Clerk JWT token. Include it in the header:\n`Authorization: Bearer <your_clerk_token>`\n\n### 📊 Rate Limiting\n- **Standard API**: 100 requests per 15 minutes.\n- **Webhooks**: 50 requests per 15 minutes.\n",
    "contact": {
      "name": "PYQDeck Support",
      "url": "https://pyqdeck.in/support",
      "email": "support@pyqdeck.in"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://backend.pyqdeck.in/api/v1",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "Forbidden": {
        "description": "The server understood the request but refuses to authorize it.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource could not be found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Detailed error message here"
          },
          "code": {
            "type": "integer",
            "example": 400
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "code": {
            "type": "integer",
            "example": 200
          }
        }
      },
...
      "Paper": {
        "type": "object",
        "required": [
          "subjectOfferingId",
          "title",
          "examYear",
          "examType",
          "slug"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "subjectOfferingId": {
            "type": "string",
            "description": "Reference to SubjectOffering"
          },
          "title": {
            "type": "string"
          },
          "examYear": {
            "type": "integer"
          },
          "examType": {
            "type": "string",
            "enum": [
              "regular",
              "re-exam",
              "supplementary",
              "end-sem",
              "internal"
            ]
          },
          "session": {
            "type": "string",
            "example": "Nov-Dec"
          },
          "regulation": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "description": "Duration in minutes"
          },
          "maxMarks": {
            "type": "number"
          },
          "slug": {
            "type": "string",
            "example": "compiler-design-2021-regular"
          },
          "uploadedBy": {
            "type": "string",
            "description": "Reference to User"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "pending",
              "approved",
              "rejected"
            ],
            "default": "pending"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Question": {
        "type": "object",
        "required": [
          "text",
          "type"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "normalizedText": {
            "type": "string",
            "description": "Cleaned text used for deduplication and search"
          },
          "type": {
            "type": "string",
            "enum": [
              "mcq",
              "short",
              "long",
              "numerical",
              "coding"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "bloomLevel": {
            "type": "string",
            "enum": [
              "remember",
              "understand",
              "apply",
              "analyze",
              "evaluate",
              "create"
            ]
          },
          "marks": {
            "type": "number"
          },
          "estimatedTime": {
            "type": "integer",
            "description": "Estimated solving time in minutes"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "References to Tag documents"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "equations": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "LaTeX equation strings"
          },
          "codeSnippets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "slug": {
            "type": "string",
            "example": "explain-lexical-analyzer-compiler-design"
          },
          "language": {
            "type": "string",
            "default": "en"
          },
          "createdBy": {
            "type": "string",
            "description": "Reference to User"
          },
          "isVerified": {
            "type": "boolean",
            "default": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Semester": {
        "type": "object",
        "required": [
          "branchId",
          "number",
          "slug"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "branchId": {
            "type": "string",
            "description": "Reference to Branch"
          },
          "number": {
            "type": "integer",
            "minimum": 1,
            "maximum": 8
          },
          "slug": {
            "type": "string",
            "example": "semester-5"
          },
          "title": {
            "type": "string",
            "example": "Semester 5"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Subject": {
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "subjectCode": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "credits": {
            "type": "number"
          },
          "slug": {
            "type": "string"
          },
          "redirectSlugs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Old slugs that 301-redirect to the current slug"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/search": {
      "get": {
        "operationId": "unifiedSearch",
        "tags": [
          "Search"
        ],
        "summary": "Global unified search across catalog entities",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search text (empty returns no-result payload)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated search results",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "questions": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Question"
                              }
                            },
                            "subjects": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Subject"
                              }
                            },
                            "papers": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Paper"
                              }
                            },
                            "totalQuestions": {
                              "type": "integer"
                            },
                            "totalSubjects": {
                              "type": "integer"
                            },
                            "totalPapers": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/branches/{branchId}/semesters": {
      "get": {
        "operationId": "listSemesters",
        "tags": [
          "Semesters"
        ],
        "summary": "List semesters for a branch",
        "parameters": [
          {
            "in": "path",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Semesters list",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Semester"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createSemester",
        "tags": [
          "Semesters"
        ],
        "summary": "Create a semester (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Payload without branchId (from URL path)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Semester"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Semester"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/branches/{branchId}/semesters/{number}": {
      "get": {
        "operationId": "getSemesterByNumber",
        "tags": [
          "Semesters"
        ],
        "summary": "Get semester by number within a branch",
        "parameters": [
          {
            "in": "path",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Semester details",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Semester"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/branches/{branchId}/semesters/{id}": {
      "patch": {
        "operationId": "updateSemester",
        "tags": [
          "Semesters"
        ],
        "summary": "Update a semester (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Semester"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Semester"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteSemester",
        "tags": [
          "Semesters"
        ],
        "summary": "Delete a semester (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}